% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_10.R
\name{clear_legal_values}
\alias{clear_legal_values}
\title{Clear Legal Values}
\usage{
clear_legal_values(.input_lines)
}
\arguments{
\item{.input_lines}{}
}
\value{

}
\description{
Function to reduce down any legal values found in the input
dataset
}
