% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_10.R
\name{find_and_score_illegal_closes}
\alias{find_and_score_illegal_closes}
\title{Find and score the illegal closes}
\usage{
find_and_score_illegal_closes(.parsed_lines)
}
\arguments{
\item{.parsed_lines}{}
}
\value{

}
\description{
Function to find and score all the illegal closes in a report
}
