% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_10.R
\name{find_and_score_incomplete_lines}
\alias{find_and_score_incomplete_lines}
\title{Find the incomplete lines and score them}
\usage{
find_and_score_incomplete_lines(.parsed_lines)
}
\arguments{
\item{.parsed_lines}{}
}
\value{

}
\description{

}
