% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_9.R
\name{find_lowpoint_risk_level}
\alias{find_lowpoint_risk_level}
\title{find the total risk level at the local minima}
\usage{
find_lowpoint_risk_level(.lava_tube_matrix)
}
\arguments{
\item{.lava_tube_matrix}{a matrix with integer values representing the height
of lava tubes}
}
\value{
a single integer value representing the total risk
}
\description{
using the matrix representation of the lava tube, determine the
total risk level, of all the local minima.
}
\details{
DETAILS
}
\section{day_9}{
NA
}

\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
