% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_9.R
\name{get_lowest_height_coords}
\alias{get_lowest_height_coords}
\title{Retrieve the lowest point coords}
\usage{
get_lowest_height_coords(.lava_tube_matrix)
}
\arguments{
\item{.lava_tube_matrix}{a matrix with integer values representing the height
of lava tubes}
}
\value{
a matrix of indexing values for to identify the local minima
}
\description{
Using the height matrix of the lava tube, find all the
coordinates of local minima
}
\details{
DETAILS
}
\section{day_9}{
NA
}

\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
