% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_9.R
\name{parse_lava_tube_matrix}
\alias{parse_lava_tube_matrix}
\title{Parse a Lava Tube Matrix}
\usage{
parse_lava_tube_matrix(.input_lines)
}
\arguments{
\item{.input_lines}{The input lines provided by AOC}
}
\value{
a matrix with integer values representing the height of lava tubes
}
\description{
Using the input lines, determine the integer matrix of the
height valuels.
}
\details{
DETAILS
}
\section{day_9}{
NA
}

\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
